# refactor: UnmarshalBy with zero-copy for string