1. Performance improved in case of `UnmarshalBy_String` .


- standard JSON codec
```
bytedance@R4374HGKKG gson % benchstat old.txt new.txt           
goos: darwin
goarch: arm64
pkg: github.com/bytedance/gg/gson
cpu: Apple M1 Pro
                                │   old.txt   │              new.txt               │
                                │   sec/op    │   sec/op     vs base               │
UnmarshalBy_Bytes-10              446.4n ± 3%   440.2n ± 2%       ~ (p=0.123 n=10)
UnmarshalBy_String-10             464.9n ± 3%   442.1n ± 4%  -4.89% (p=0.000 n=10)
UnmarshalBy_LargeJSON_String-10   12.55µ ± 2%   12.36µ ± 3%  -1.56% (p=0.035 n=10)
UnmarshalBy_LargeJSON_Bytes-10    12.68µ ± 4%   12.36µ ± 2%  -2.54% (p=0.002 n=10)
geomean                           2.397µ        2.335µ       -2.61%
```
- sonic.ConfigFastest
```
bytedance@R4374HGKKG gson % benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/bytedance/gg/gson
cpu: Apple M1 Pro
                                │   old.txt   │              new.txt               │
                                │   sec/op    │   sec/op     vs base               │
UnmarshalBy_Bytes-10              252.3n ± 3%   247.1n ± 0%       ~ (p=0.516 n=10)
UnmarshalBy_String-10             269.5n ± 4%   251.1n ± 2%  -6.85% (p=0.000 n=10)
UnmarshalBy_LargeJSON_String-10   3.586µ ± 3%   3.337µ ± 2%  -6.94% (p=0.000 n=10)
UnmarshalBy_LargeJSON_Bytes-10    3.539µ ± 1%   3.376µ ± 3%  -4.59% (p=0.000 n=10)
geomean                           963.7n        914.3n       -5.13%

```

2. Updated docs as well.